requirements.txt
flask==2.3.3
chromadb==0.4.15
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
pandas==2.0.3
pyarrow==14.0.1